Cleanups and fixups post PR

+++ src/test/java/com/zepben/blobstore/sqlite/SqliteBlobStoreTest.java
@@ -37,21 +37,12 @@ import static org.junit.jupiter.api.Assertions.*;
 @SuppressWarnings("SqlNoDataSourceInspection")
 public class SqliteBlobStoreTest {
 
-    private static final String tempDB = "store.db";
     private static Path dbFile;
-
-    static {
-        try {
-            dbFile = Files.createTempFile(Path.of("/tmp"), tempDB, "");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
 
     @BeforeEach
     public void before() throws IOException {
-        Files.deleteIfExists(dbFile);
-        dbFile = Files.createTempFile(Path.of("/tmp"), tempDB, "");
+        if (dbFile != null) Files.deleteIfExists(dbFile);
+        dbFile = Files.createTempFile(Path.of(System.getProperty("java.io.tmpdir")), "store.db", "");
     }
 
     @AfterAll
